rename seedlist to inboxPlacement in nav items
